+++ src/UI/TransactionInfo/components/transactionInfoFields/TransactionInfoMethod/TransactionInfoMethod.tsx
@@ -3,11 +3,17 @@ import React from 'react';
 import { TransactionActionCategoryEnum } from 'types/serverTransactions.types';
 import { getTransactionMethod } from 'utils/transactions/getInterpretedTransaction/helpers/getTransactionMethod';
 
+import {
+  WithClassnameType,
+  WithTransactionType
+} from '../../../../../UI/types';
 import { DetailItem } from '../../DetailItem';
-import { WithTransactionType } from '../../../../../UI/types';
 import { TransactionAction } from '../../TransactionAction';
 
-export const TransactionInfoMethod = ({ transaction }: WithTransactionType) => {
+export const TransactionInfoMethod = ({
+  className,
+  transaction
+}: WithTransactionType & WithClassnameType) => {
   const showMethod = transaction.action && transaction.action.category;
   const showAction =
     transaction.action?.category !== TransactionActionCategoryEnum.scCall;
@@ -18,12 +24,12 @@ export const TransactionInfoMethod = ({ transaction }: WithTransactionType) => {
 
   return (
     <>
-      <DetailItem title='Method'>
+      <DetailItem className={className} title='Method'>
         {getTransactionMethod(transaction)}
       </DetailItem>
 
       {showAction && (
-        <DetailItem title='Transaction Action'>
+        <DetailItem className={className} title='Transaction Action'>
           <TransactionAction transaction={transaction} />
         </DetailItem>
       )}